docs(button): optimize basic usage demo
