+++ examples/sites/demos/app/button/base-composition-api.vue
@@ -24,11 +24,27 @@
       <tiny-button type="warning" round> 警告按钮 </tiny-button>
       <tiny-button type="danger" round> 危险按钮 </tiny-button>
     </tiny-row>
+    <tiny-row>
+      <tiny-button :icon="IconSearch" circle></tiny-button>
+      <tiny-button type="primary" :icon="IconEdit" circle></tiny-button>
+      <tiny-button type="success" :icon="IconYes" circle></tiny-button>
+      <tiny-button type="info" :icon="IconMail" circle></tiny-button>
+      <tiny-button type="warning" :icon="IconStarO" circle></tiny-button>
+      <tiny-button type="danger" :icon="IconDel" circle></tiny-button>
+    </tiny-row>
   </tiny-layout>
 </template>
 
 <script setup>
 import { Button as TinyButton, Layout as TinyLayout, Row as TinyRow } from '@opentiny/vue'
+import { iconDel, iconYes, iconEdit, iconMail, iconStarO, iconSearch } from '@opentiny/vue-icon'
+
+const IconDel = iconDel()
+const IconYes = iconYes()
+const IconEdit = iconEdit()
+const IconMail = iconMail()
+const IconStarO = iconStarO()
+const IconSearch = iconSearch()
 </script>
 
 <style scoped>
